+++ lib/src/metrics/class/_default_class_metrics.dart
@@ -1,15 +1,19 @@
+import 'cbo.dart';
 import 'class_length.dart';
 import 'class_metric.dart';
 import 'dit.dart';
 import 'lcom4.dart';
 import 'noc.dart';
 import 'nom.dart';
+import 'rfc.dart';
 import 'wmc.dart';
 
 const List<ClassMetric> defaultClassMetrics = [
   NumberOfMethods(),
   WeightedMethodsPerClass(),
   Lcom4(),
+  CouplingBetweenObjects(),
+  ResponseForClass(),
   DepthOfInheritanceTree(),
   NumberOfChildren(),
   ClassLength(),