+++ lib/src/metrics/library/_default_library_metrics.dart
@@ -1,0 +1,10 @@
+import 'coupling.dart';
+import 'library_metric.dart';
+
+const List<LibraryMetric> defaultLibraryMetrics = [
+  EfferentCoupling(),
+  AfferentCoupling(),
+  Instability(),
+  Abstractness(),
+  DistanceFromMainSequence(),
+];